requirements.txt
# ==========================================
# SecureHospital AI - Requirements
# ==========================================
# Your existing requirements + production additions

aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.1
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
bidict==0.23.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
docutils==0.22.3
Faker==37.12.0
fastapi==0.121.1
frozenlist==1.8.0
h11==0.16.0
htmx==0.0.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
markdown2==2.5.4
MarkupSafe==3.0.3
multidict==6.7.0
nicegui==3.2.0
openai==2.7.2
orjson==3.11.4
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.3
PyYAML==6.0.3
requests==2.32.5
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.2.0
yarl==1.22.0

# ==========================================
# Production additions (for Vercel)
# ==========================================
whitenoise==6.7.0
gunicorn==22.0.0